Move :prefs modules out of :common
