+++ app-tv/build.gradle.kts
@@ -94,7 +94,6 @@ slack {
 }
 
 dependencies {
-    implementation(projects.common.prefs.impl)
     implementation(projects.common.translations)
     implementation(projects.libraries.foundation.coroutines)
     implementation(projects.libraries.lessons.api)
@@ -103,6 +102,7 @@ dependencies {
     implementation(projects.libraries.media.service)
     implementation(projects.services.lessons.impl)
     implementation(projects.services.media.impl)
+    implementation(projects.services.prefs.impl)
     implementation(projects.services.storage.impl)
 
     implementation(libs.androidx.activity.compose)
