+++ app/build.gradle.kts
@@ -137,7 +137,6 @@ dependencies {
     implementation(projects.common.designCompose)
     implementation(projects.common.lessonsData)
     implementation(projects.common.network)
-    implementation(projects.common.prefs.impl)
     implementation(projects.common.runtimePermissions)
     implementation(projects.common.translations)
     implementation(projects.features.appWidgets)
@@ -152,6 +151,7 @@ dependencies {
     implementation(projects.services.circuit.impl)
     implementation(projects.services.lessons.impl)
     implementation(projects.services.media.impl)
+    implementation(projects.services.prefs.impl)
     implementation(projects.services.storage.impl)
     implementation(projects.services.workers.impl)
