+++ common/core/build.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023. Adventech <info@adventech.io>
+ * Copyright (c) 2024. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -41,7 +41,6 @@ android {
 
 dependencies {
     api(projects.common.misc)
-    api(projects.common.prefs.api)
     api(libs.circuit.runtime)
     implementation(libs.androidx.appcompat)
     implementation(libs.androidx.browser)
