+++ common/lessons-data/build.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023. Adventech <info@adventech.io>
+ * Copyright (c) 2024. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -43,10 +43,10 @@ dependencies {
     api(projects.common.models)
     implementation(projects.common.misc)
     implementation(projects.common.network)
-    implementation(projects.common.prefs.api)
     implementation(projects.libraries.foundation.android)
     implementation(projects.libraries.foundation.coroutines)
     implementation(projects.libraries.lessons.api)
+    implementation(projects.libraries.prefs.api)
     implementation(projects.libraries.storage.api)
 
     implementation(libs.google.hilt.android)
