+++ features/bible/build.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023. Adventech <info@adventech.io>
+ * Copyright (c) 2024. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -51,6 +51,7 @@ dependencies {
     implementation(projects.common.translations)
     implementation(projects.common.lessonsData)
     implementation(projects.features.reader)
+    implementation(projects.libraries.prefs.api)
     implementation(projects.libraries.foundation.coroutines)
 
     implementation(libs.google.material)
